webapp: revise presentation of "ignore SoC" switch
in order to avoid a negation, the switch is re-labelled to "use voltage
thresholds only". the hint text explains that SoC thresholds can be
configured once the switch is turned off.

this also allows to move the switch to the voltage thresholds card, so
the whole SoC thresholds card can be removed if the switch is enabled.

as the usage of SoC thresholds is not recommended in general, the
default value for "ignore SoC" is changed to true.